Extract view more details
